+++ src/validators/loan.schema.js
@@ -1,8 +1,8 @@
 import { z } from "zod";
 
 export const borrowSchema = z.object({
-  bookCopyId: z.number(),
-  days: z.number().min(1).max(30) // max 30 hari
+  bookCopyId: z.number().int(),
+  days: z.number().int().min(1).max(30) // max 30 hari
 });     
 
 export const returnSchema = z.object({
